Added a functional ini-style configuration file parser: ConfigFile. It will be used to read configuration files for planetary systems and saved missions. The syntax is flexible and human-readable. Users (including non-technical users) should be able to produce descriptions for missions and planetary systems with little difficulty.
ConfigParser is somewhat content-neutral. It does not construct Planet or (future) Mission objects. Object factories will do this by calling the appropriate getToken(string), getDistance(string), and other such functions which will search for a property of a certain name and interpret it. For example, all planets should have a property called "sma" (for semi-major axis). The Planet factory will call getDistance("sma"). If the value entered by the user is "4e24 kg" then ConfigParser will interpret the scientific notation and units. Thus, ConfigParser does not know what a string is supposed to be until an object factory tells it how to interpret it.

There is still much work to be done on ConfigParser. It should be much more error-tolerant, throwing whatever C++ uses for exceptions whenever it runs into an error, and also give the user an accurate impression of where the error occured. It should also support comments. I also haven't run anything through Valgrind, so memory management is probably a mess. Most importantly, it doesn't actually have the full API needed by object factories.

But for now, everything builds and works.